Move CantoneseUtils out of ChineseUtils
