+++ src/jyut-dict/logic/entry/entryspeaker.cpp
@@ -1,4 +1,5 @@
 #include "entryspeaker.h"
+#include "logic/utils/cantoneseutils.h"
 #include "logic/utils/chineseutils.h"
 
 #include <QCoreApplication>
@@ -204,7 +205,7 @@ int EntrySpeaker::speak(const QLocale::Language &language,
 
         std::vector<std::string> syllables;
         if (language == QLocale::Cantonese || country == QLocale::HongKong) {
-            ChineseUtils::segmentJyutping(string, syllables);
+            CantoneseUtils::segmentJyutping(string, syllables);
         } else {
             QString mutableString = string;
             mutableString.replace("u:", "ü");
